--- accounts/getters/getAccounts.js
@@ -1,3 +1,0 @@
-import { prop } from 'ramda'
-
-export default prop('accounts')
