+++ actions.js
@@ -1,3 +1,3 @@
 export { actions as agents } from './agents'
-export { actions as accounts } from './accounts'
+export { actions as credentials } from './credentials'
 export { actions as authentication } from './authentication'
