+++ authentication/getters/getRegisterProps.js
@@ -1,0 +1,3 @@
+import { createStructuredSelector } from 'reselect'
+
+export default createStructuredSelector({})
