+++ authentication/hoc/userIsAuthenticated.js
@@ -1,10 +1,11 @@
 import { connectedRouterRedirect } from 'redux-auth-wrapper/lib/history4/redirect'
 import { propOr } from 'ramda'
-import { getAccount } from '../getters'
+
+import getCredential from '../getters/getCredential'
 
 export default connectedRouterRedirect({
   wrapperDisplayName: 'UserIsAuthenticated',
-  authSelector: getAccount,
+  authSelector: getCredential,
   predicate: propOr(null, 'accessToken'),
   redirectPath: '/sign-in',
   allowRedirectBack: true
