+++ authentication/hoc/userIsAuthenticatedOrHome.js
@@ -2,13 +2,13 @@ import locationHelperBuilder from 'redux-auth-wrapper/lib/history4/locationHelpe
 import { propOr } from 'ramda'
 import { connectedRouterRedirect } from 'redux-auth-wrapper/lib/history4/redirect'
 
-import { getAccount } from '../getters'
+import getCredential from '../getters/getCredential'
 
 const locationHelper = locationHelperBuilder({})
 
 export default connectedRouterRedirect({
   wrapperDisplayName: 'UserIsAuthenticatedOrHome',
-  authSelector: getAccount,
+  authSelector: getCredential,
   predicate: propOr(null, 'accessToken'),
   redirectPath: '/',
   allowRedirectBack: false
