+++ credentials/getters/getCredentials.js
@@ -1,0 +1,3 @@
+import { prop } from 'ramda'
+
+export default prop('credentials')
