+++ epic.js
@@ -1,11 +1,11 @@
 const { combineEpics } = require('redux-observable')
 
 import { epic as agents } from './agents'
-import { epic as accounts } from './accounts'
+import { epic as credentials } from './credentials'
 import { epic as authentication } from './authentication'
 
 export default combineEpics(
   agents,
-  accounts,
+  credentials,
   authentication
 )
