+++ getters.js
@@ -1,7 +1,7 @@
 export { default as getAgents } from './agents/getters/getAgents'
-export { default as getAccounts } from './accounts/getters/getAccounts'
+export { default as getCredentials } from './credentials/getters/getCredentials'
 export { default as getAuthentication } from './authentication/getters/getAuthentication'
-export { default as getAccount } from './authentication/getters/getAccount'
+export { default as getCredential } from './authentication/getters/getCredential'
 export { default as getAccessToken } from './authentication/getters/getAccessToken'
 export { default as getSigningIn } from './authentication/getters/getSigningIn'
 export { default as getIsAuthenticated } from './authentication/getters/getIsAuthenticated'
